Managed to write a script called reset_inventory.py to zero out invenotory, tally up total count per part in csv file, updated inventory to new total if it was in DB, and if it was never in the DB for inventories, then to create new inventory for it.
Now all inventory is added except for a few parts that I don't know
what model nor repair they go to. The script will list out any parts not
in DB's parts table. It will add csv: "monthly_inventory.csv", which is a
a hardlinked file from another directory, then it adds the count to each part
in the DB.