Added location_desc to inventory object when doing HTTP GET to the get_inventory api route.
Did this so I could show the name of the location for each part, instead of the
location_id, which deosn't help us know where parts are in human speak.

Also, had to fix this up on the front end to show once per inventory object,
instead of twice, once per the object keys in each inventory object due
to me adding one more property to each object.