Created the Register route to create new users easily. Created the requirment for login to use this app. Tokens are 8 hours long. Tested the register route. Also, after registering, you log in right away and get redirected to "/". Not the best way, but will be able to get my users added really fast.
Put in placeholder text for add_model and add_repair input feilds.
For adding new repair, the model name is added to beginning automatically.
Only the repair name is needed without any "-" characters.

Need to test login and registering more.

In Inventory_Management:
Need to elimate the location input feild for adding inventory. This
should come automatically from the location of the current inventory.

After the "use one" button is pressed in Inventory_Management, need the inventory to
automatically upgrade. Will require another api call and DB query.

In Recieve Parts:
Need example of the csv file to upload.

Need feebback of successful inventory update or an error message.

Need to fix the "/" in the receive parts route. URL encoding or decoding error.